rename: removed s from model names
