+++ backend/controllers/user.controller.js
@@ -1,4 +1,4 @@
-import User from "../models/users.model.js"
+import User from "../models/user.model.js"
 import bcrypt from "bcryptjs"
 import jwt from "jsonwebtoken"
